HANDLE FREQ DEPENDING ON FORMAT IN CONFIGURATION
IEEE Std C37.118.2-2011 Table 6 states FREQ is the frequency deviation, unless 32bit floating pt.
In the case of 16bit float it's the deviation and we unpack and restore nominal offset.
In the case of 32bit float it's the actual value and we unpack without adjustment.
Also not necessary to restrict to +/- 32.767Hz as it is with 16bit integer.