Frequency settings matched to standard #4
When FREQ is 16-bit integer, it is considered to be frequency deviation from nominal (in mHz) and it is added to nominal frequency value to get actual value of frequency (see one of my commits) , divided by 1000.
When FREQ is floating points, it is actual frequency value that should be written in Hz (because range is -32767 <= freq<=32767, so max value would be 32.767Hz if we use mHz for input).

This is only my understanding of standard and should be checked by others.